ACT: Further changes to back-tracking
It is now possible to set the number of times that a back-tracking event must be initiated at (or below) a given streamline length before the length of truncation is increased.

+++ src/dwi/tractography/ACT/method.h
@@ -155,7 +155,7 @@ namespace MR
             bool in_pathology() const { return (tissue_values.valid() && tissue_values.is_path()); }
 
 
-            unsigned int sgm_depth;
+            size_t sgm_depth;
             bool seed_in_sgm;
             bool sgm_seed_to_wm;
